refactor: errorMessage 함수 위치 변경
